+++ Assets/Scripts/Tower/SingleTargetTower.cs
@@ -8,7 +8,7 @@ public class SingleTargetTower : Tower {
     [SerializeField] private SphereCollider targetCollider;
     [SerializeField] private int steps;
     [SerializeField] private float range;
-    [SerializeField] private int damage;
+    [SerializeField] private float damage;
     [SerializeField] private int cost;
     [SerializeField] private float fireRate;
     [SerializeField] private float drawHeight;
@@ -56,7 +56,7 @@ public class SingleTargetTower : Tower {
         }
     }
 
-    protected override int Damage {
+    protected override float Damage {
         get {
             return damage;
         }
@@ -101,7 +101,7 @@ public class SingleTargetTower : Tower {
         while (true) {
             if (targets.Count > 0) {
                 if (targets[0].isAlive()) {
-                    targets[0].takeDamage(damage);
+                    targets[0].takeDamage((int)damage);
                 } else {
                     targets.RemoveAt(0);
                 }
